+++ tests/dummy/app/locales/en/translations.js
@@ -1125,34 +1125,6 @@ $.extend(true, translations, {
           'title': ''
         }
       },
-      'flexberry-simpleolv': {
-        'limit-function-example': {
-          'caption': 'Flexberry-simpleolv. Limit function example',
-          'captionFirstLimitFunction': 'Limit function №1',
-          'captionSecondLimitFunction': 'Limit function №2',
-          'captionClearLimitFunction': 'Clear limit function'
-        },
-        'settings-example': {
-          'caption': 'Flexberry-simpleolv. Settings example'
-        },
-        'toolbar-custom-buttons-example': {
-          'caption': 'Flexberry-simpleolv. Custom buttons on toolbar',
-          'custom-message': 'Hello!',
-          'custom-button-name': 'Send hello'
-        },
-        'on-edit-form': {
-          'caption': 'Flexberry-simpleolv. FlexberryObjectlistview custom data sample'
-        },
-        'custom-filter': {
-          'caption': 'Flexberry-simpleolv. Custom filter'
-        },
-        'configurate-rows': {
-          'caption': 'Flexberry-simpleolv. Configurate rows'
-        },
-        'selected-rows': {
-          'caption': 'Flexberry-simpleolv. Setected rows'
-        }
-      },
       'flexberry-simpledatetime': {
         'settings-example': {
           'caption': 'Flexberry-simpledatetime. Settings example'