+++ tests/dummy/app/locales/ru/translations.js
@@ -1146,34 +1146,6 @@ $.extend(true, translations, {
           'title': ''
         }
       },
-      'flexberry-simpleolv': {
-        'limit-function-example': {
-          'caption': 'Flexberry-simpleolv. Функция ограничения',
-          'captionFirstLimitFunction': 'Функция ограничения №1',
-          'captionSecondLimitFunction': 'Функция ограничения №2',
-          'captionClearLimitFunction': 'Очистить функцию ограничения'
-        },
-        'settings-example': {
-          'caption': 'Flexberry-simpleolv. Пример работы с настройками'
-        },
-        'toolbar-custom-buttons-example': {
-          'caption': 'Flexberry-simpleolv. Пользовательские кнопки',
-          'custom-message': 'Привет!',
-          'custom-button-name': 'Передать привет'
-        },
-        'on-edit-form': {
-          'caption': 'Flexberry-simpleolv. FlexberryObjectlistview пример произвольных данных'
-        },
-        'custom-filter': {
-          'caption': 'Flexberry-simpleolv. Настройка фильтра'
-        },
-        'configurate-rows': {
-          'caption': 'Flexberry-simpleolv. Раскраска строк'
-        },
-        'selected-rows': {
-          'caption': 'Flexberry-simpleolv. Выбранные строки'
-        }
-      },
       'flexberry-simpledatetime': {
         'settings-example': {
           'caption': 'Flexberry-simpledatetime. Пример работы с настройками'